feat:search course

+++ src/main/java/com/cmsbackend/controller/user_controller/user_vo/AddRequest.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @ApiModel
 public class AddRequest {
     String Account;
-    Integer Identity;
+    Long Identity;
     String Password;
     String Name;
 }
